requirements.txt
flask
flask-sqlalchemy
flask-wtf
python-dotenv
openai